Add psuedocode sketch of bit_flip_CBC function
I added a psuedocode sketch of a function to do the bit_flipping to produce
the intended target bytes in a given block of the decrypted cipher. It should
work fine but I need to do a quick xor of the target vs the known plaintext
in the target block. Will do that this afternoon.